+++ tests/test_stat_ecdf.py
@@ -16,8 +16,7 @@ def test_computed_y_column():
     p = (
         ggplot(data, aes("x"))
         + stat_ecdf(size=2)
-        # Should be able to used computed y column & create a
-        # new mapped column also called y
-        + stat_ecdf(aes(y=after_stat("y-0.2")), size=2, color="blue")
+        # Should be able to use calculated & create a new column called y
+        + stat_ecdf(aes(y=after_stat("ecdf-0.2")), size=2, color="blue")
     )
     assert p == "computed_y_column"